+++ src/helpers/removeUrlAtEnd.ts
@@ -4,7 +4,7 @@
  * @param text The text to get URLs from.
  * @returns An array of URLs.
  */
-export function removeUrlAtEnd(url?: string, content?: string): string {
+export function removeUrlAtEnd(url: string, content: string): string {
     if (url && content) {
         content = content.trimEnd();
         const indexOfUrl = content.indexOf(url);
